requirements.txt
tensorflow
matplotlib
pillow
fastapi
python-multipart
